Allows handling load/save errors
When an error occurs during loading/saving of the app state, it is
now returned instead of just an error being printed. This way the
error can be handled where it is needed. To facilitate this,
`loadAppState` and `saveAppState` return a `Result`. In the case of
saving, it returns the save date so that this may be displayed.
This information should of course not be persisted itself, so a new
state section "meta" was introduced, that is not persisted.