Allows opening braces on different line
Normally we don't want them to be on a different line, but in
if-statements with multiple lines in the condition, it is sometimes
more clear to have the brace on a separate line. That is in fact the
only situation where this should be done, and in that case the brace
has to be the only single character on that line, with indentation
equal to that of the `if` keyword.